# api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReflectanceSurfacesResponse": {
        "description": "Response from GET /session/room/surfaces.",
        "properties": {
          "surfaces": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SurfaceInfo"
            },
            "title": "Surfaces",
            "type": "object"
          }
        },
        "required": [
          "surfaces"
        ],
        "title": "ReflectanceSurfacesResponse",
        "type": "object"
      },
...
      "RoomUpdateResponse": {
        "description": "Response from PATCH /session/room.",
        "properties": {
          "message": {
            "default": "Operation completed successfully",
            "title": "Message",
            "type": "string"
          },
          "room": {
            "$ref": "#/components/schemas/RoomGeometry"
          },
          "state_hashes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Hashes"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "room"
        ],
        "title": "RoomUpdateResponse",
        "type": "object"
      },
...
      "SessionInitRequest": {
        "description": "Request to initialize a session with full project state",
        "properties": {
          "lamps": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SessionLampInput"
            },
            "title": "Lamps",
            "type": "array"
          },
          "room": {
            "$ref": "#/components/schemas/SessionRoomConfig"
          },
          "zones": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SessionZoneInput"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "room"
        ],
        "title": "SessionInitRequest",
        "type": "object"
      },
      "SessionInitResponse": {
        "description": "Response after session initialization",
        "properties": {
          "lamp_count": {
            "title": "Lamp Count",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "zone_count": {
            "title": "Zone Count",
            "type": "integer"
          }
        },
        "required": [
          "success",
          "message",
          "lamp_count",
          "zone_count"
        ],
        "title": "SessionInitResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API for UV disinfection simulation",
    "title": "Illuminate API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/session/init": {
      "post": {
        "description": "Initialize a new session Room from frontend state.\n\nThis creates the source of truth Room for this session that will be mutated\nby subsequent PATCH/POST/DELETE calls. Any existing Room in this session is replaced.\n\nRequires X-Session-ID header.",
        "operationId": "init_session_api_v1_session_init_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Session-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Init Session",
        "tags": [
          "Session",
          "Session"
        ]
      }
    },
...
    "/api/v1/session/room": {
      "patch": {
        "description": "Update room configuration properties.\n\nOnly provided fields are updated. Room dimensions, units, and other\nsettings can be changed without recreating the entire Room. The response\nechoes the resulting room geometry (shape, vertices, wall ids) and the\nper-surface reflectance state, which guv_calcs carries across shape\nchanges by edge index.\n\nRequires X-Session-ID header.",
        "operationId": "update_session_room_api_v1_session_room_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Session-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRoomUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Session Room",
        "tags": [
          "Session",
          "Session"
        ]
      }
    },
    "/api/v1/session/room/surfaces": {
      "get": {
        "description": "Return per-surface reflectance grid info (spacing and num_points).\n\nUsed by the ReflectanceSettingsModal to show the backend's actual\nsurface resolution rather than frontend defaults.\n\nRequires X-Session-ID header.",
        "operationId": "get_room_surfaces_api_v1_session_room_surfaces_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Session-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReflectanceSurfacesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Room Surfaces",
        "tags": [
          "Session",
          "Session"
        ]
      }
    },
...
    "/api/v1/session/state-hashes": {
      "get": {
        "description": "Return current state hashes from room.get_calc_state() and room.get_update_state().\n\nThe frontend uses these to detect when parameters have changed since the last\ncalculation, enabling granular staleness detection per-zone.\n\nRequires X-Session-ID header.",
        "operationId": "get_state_hashes_api_v1_session_state_hashes_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Session-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateHashesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get State Hashes",
        "tags": [
          "Session",
          "Session"
        ]
      }
    },
    "/api/v1/session/status": {
      "get": {
        "description": "Get current session status for debugging.\n\nRequires X-Session-ID header.",
        "operationId": "get_session_status_api_v1_session_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Session-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session Status",
        "tags": [
          "Session",
          "Session"
        ]
      }
    },
...
    "/api/v1/session/units": {
      "patch": {
        "description": "Change the unit system for the session.\n\nCalls room.set_units() which permanently converts all dimensions\n(room, lamps, zones) to the new unit system.\n\nReturns all converted coordinates so the frontend can update its stores.\n\nRequires X-Session-ID header.",
        "operationId": "set_session_units_api_v1_session_units_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Session-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUnitsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetUnitsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Session Units",
        "tags": [
          "Session",
          "Session"
        ]
      }
    },
...
  }
}